design outline of strategies and factory.
Got rid of configuring task ranker with config.yml file. Instead
using options as this repository is to be included as a sub-module
and not run standalone.
Added a dummy strategy for ranking tasks based on cpu-shares allocation.
Added a factory to instantiate a strategy to be then built.

Strategy now stores a callback object called receiver. Each time
the tasks are ranked, the results are passed through the callback
function Receive(...).

Added support for go modules.
Updated README.md to be consistent with the above changes.
Updated test code.