label matchers + operators + query builder.
Label matchers, metrics and range (time unit and duration) information
is dependent on the strategy. Therefore, instead of DataFetcher holding
those members, the strategy now holds them. Updated the strategy API
with necessary getters and setters.

The data fetcher now composes a reference to the strategy. This
reference is used to get the required pieces of information to build
the query string such as metric name, label matchers, etc.

Enumerated time units that implements Stringer. On calling String()
we get the abbreviation of the unit as used in the query string.
For example, Seconds.String() would return 's' and Minutes.String()
would return 'm'.

Enumerated label match operators. LabelMatcher now implements Stringer.
On calling String() we get a comma separated list of label matchers.
For example, if label key is 'test', operator is '=' and value is 'v'
we get 'test="v"'.

Currently, as a placeholder, the data fetcher uses the query builder to
build the query string and returns it. The task ranker currently just
feeds that query string to the strategy and the strategy just echoes it.

Added test code for query builder.

Updated the test code for TaskRanker.

Updated readme.