dedicated label matchers.
Added dedicated label matchers that can be used filter prometheus data
based on a task's unique identifier and/or name of the host on which
a task is running.
Strategies can mandate the use of dedicated label matchers, especially
if tasks' runtime information is used for ranking.
LabelMatcher not contains a type that can be set to one of the following.
1. Other (default).
2. TaskID - flag label as one that can be used to fetch unique identifier
	of task.
3. TaskHostname - flag labe as one that can be used to fetch name of host
	on which task is running.

Updated README to include instructions on providing dedicated labels.