cpu util (%) based ranking of colocated tasks.
Added a new task ranking strategy that ranks colocated tasks based
on their cpu utilization (%) in the past 5 intervals of time.
The prometheus scrape interval (in seconds) is used as an interval of time.

The metric used is container_cpu_usage_seconds_total that provides
the cpu usage in seconds on each cpu for each task. The cpu utilization
is calculated as shown below.
cpu_util(%) = 100 * ((cpu_usage[n-1] - cpu_usage[0]) / ((n-1) * scrape_interval))

where,
1. n is the number of data points fetched.
2. (n-1) * scrape_interval gives us the elapsed time.
3. (cpu_usage[n-1] - cpu_usage[0]) gives us the #seconds utilized.

This strategy uses the dedicated label of type TaskHostname to
retrieve host information to group colocated tasks.
This strategy also uses the TaskID dedicated label to retrieve
the task identifier to match cpu usage information to a task.