Release 0.5.0 (#17)
configurable range + strategy (#16)

configure strategy + range duration with options.

With this commit, the time duration for range queries can also be
configured. This allows for more fine grained control over the number
of data points on which the strategies is to be applied.

For backwards compatibility the earlier WithStrategy(...) function
can still be used. However, note that this results in the default
setting of range duration (as exposed by the strategy) being used.
If the time duration needs to be configured, then WithStrategyOptions(...)
should now be used.

Retrofitted cpushares task ranking strategy to now parse either a matrix
or a vector depending on whether a range query is used.

Fixed the test code.

Added sample code to readme.