feature: multiple metrics per query
with this commit, we can now have ranking strategies that require
data for multiple metrics.
Using the query format {__name__="m1|m2|...",...} where m1, m2 etc
are the names of the metrics.

Updated the unit tests to adhere to the changes.

No change was made to the end-to-end tests and ranker_test.go.
This change is backwards compatible.